pyproject.toml
[project]
name = "customer-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Niels", email = "niels.scholz01@web.de" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "llama-index-core>=0.12.0",
    "llama-index-llms-ollama>=0.5.0",
    "llama-index-embeddings-ollama>=0.5.0",
    "llama-index-vector-stores-chroma>=0.4.0",
    "chromadb>=0.5.0",
    "langfuse>=2.58.0",
    "openinference-instrumentation-llama-index>=3.0.0",
]

[project.scripts]
customer-bot = "customer_bot:main"
customer-bot-api = "customer_bot.api.main:main"
customer-bot-ingest = "customer_bot.ingest.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "pytest-xdist>=3.6",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
xfail_strict = true
markers = [
    "unit: deterministic tests with no external services",
    "integration: multi-component tests",
    "slow: slower tests excluded from default loops",
    "network: tests requiring network or external endpoints",
]
filterwarnings = [
    "error::DeprecationWarning",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
